+++ src/components/autocompleteInput.js
@@ -90,6 +90,7 @@
     const initalValue = defaultValue.replace(/\n/g, '');
     const [value, setValue] = useState(initalValue);
     const [debouncedInputValue, setDebouncedInputValue] = useState();
+    const [debouncedCurrentValue, setDebouncedCurrentValue] = useState();
     const [interactionFilter, setInteractionFilter] = useState({});
     const defaultValueEvaluatedRef = useRef(false);
 
@@ -646,6 +647,32 @@
 
     const currentValue = getValue();
 
+    useEffect(() => {
+      if (currentValue !== debouncedCurrentValue) {
+        setTimeout(() => {
+          setDebouncedCurrentValue(currentValue);
+        }, 250);
+      } else {
+        let triggerEventValue = '';
+
+        if (value) {
+          triggerEventValue = !isListProperty ? value[valueProp.name] : value;
+        }
+        changeContext.current = { modelData: value };
+        B.triggerEvent('onChange', triggerEventValue, changeContext.current);
+      }
+    }, [currentValue]);
+
+    // In the first render we want to make sure to convert the default value
+    if (!inputValue && currentValue) {
+      setValue(currentValue);
+      if (isListProperty) {
+        setInputValue(currentValue);
+      } else {
+        setInputValue(currentValue[searchProp.name].toString());
+      }
+    }
+
     const renderLabel = (option) => {
       let optionLabel = '';
       const emptyPropertyPath =
@@ -706,22 +733,7 @@
           })}
           onChange={(_, newValue) => {
             setValue(newValue || '');
-
-            let triggerEventValue;
-
-            if (!isListProperty) {
-              triggerEventValue = currentValue
-                ? currentValue[valueProp.name]
-                : '';
-            } else if (isListProperty) {
-              triggerEventValue = currentValue || '';
-            }
-            changeContext.current = { modelData: newValue };
-            B.triggerEvent(
-              'onChange',
-              triggerEventValue,
-              changeContext.current,
-            );
+            setDebouncedCurrentValue(newValue);
           }}
           onInputChange={(event, newValue) => {
             let validation = event ? event.target.validity : null;
@@ -755,7 +767,7 @@
                   type="hidden"
                   key={value[valueProp.name] ? 'hasValue' : 'isEmpty'}
                   name={nameAttribute || name}
-                  value={getHiddenValue(currentValue)}
+                  value={getHiddenValue(debouncedCurrentValue)}
                 />
               )}
               <TextField